rename package pkg/discovery to satellite/discovery
