+++ satellite/discovery/service_test.go
